Tarea de clases acabada

+++ app/src/main/java/codigo/Ejercicio1/CambioMoneda/Banco.java
@@ -5,13 +5,13 @@ public class Banco {
     public Banco(){
         System.out.println("Inicializando Banco");
     }
-    public double exchangeBolivianos(int amount, String type){
+    public double exchangeBolivianos(double amount, String type){
         return Empresa.getCambio().convertToBoliviano(amount, type);
     }
-    public double exchangeDollars(int amount, String type){
+    public double exchangeDollars(double amount, String type){
         return Empresa.getCambio().convertToDollar(amount, type);
     }
-    public double exchangeEuros(int amount, String type){
+    public double exchangeEuros(double amount, String type){
         return Empresa.getCambio().convertToEuro(amount, type);
     }
