+++ app/src/main/java/codigo/Ejercicio1/CambioMoneda/CasaDeCambio.java
@@ -5,13 +5,13 @@ public class CasaDeCambio {
     public CasaDeCambio(){
         System.out.println("Inicializando CasaDeCambio");
     }
-    public double exchangeBolivianos(int amount, String type){
+    public double exchangeBolivianos(double amount, String type){
         return Empresa.getCambio().convertToBoliviano(amount, type);
     }
-    public double exchangeDollars(int amount, String type){
+    public double exchangeDollars(double amount, String type){
         return Empresa.getCambio().convertToDollar(amount, type);
     }
-    public double exchangeEuros(int amount, String type){
+    public double exchangeEuros(double amount, String type){
         return Empresa.getCambio().convertToEuro(amount, type);
     }
     public void depositarDinero(double amount){
